+++ apps/backend/prisma/seed.ts
@@ -215,7 +215,7 @@ async function seed() {
       data: {
         id: "team_admin",
         description: "2",
-        contained_permission_ids: ["$read_members", "$update_team"],
+        contained_permission_ids: ["$read_members", "$remove_members", "$update_team"],
       }
     }
   );