+++ apps/dashboard/src/components/payments/item-dialog.tsx
@@ -4,7 +4,6 @@ import { FormDialog } from "@/components/form-dialog";
 import { InputField, SelectField } from "@/components/form-fields";
 import { AdminProject } from "@stackframe/stack";
 import { branchPaymentsSchema } from "@stackframe/stack-shared/dist/config/schema";
-import { toast } from "@stackframe/stack-ui";
 import { userSpecifiedIdSchema } from "@stackframe/stack-shared/dist/schema-fields";
 import { has } from "@stackframe/stack-shared/dist/utils/objects";
 import * as yup from "yup";