+++ packages/stack-shared/src/interface/client-interface.ts
@@ -14,6 +14,7 @@ import { AuthenticationResponseJSON, PublicKeyCredentialCreationOptionsJSON, Pub
 import { wait } from '../utils/promises';
 import { Result } from "../utils/results";
 import { deindent } from '../utils/strings';
+import { urlString } from '../utils/urls';
 import { ConnectedAccountAccessTokenCrud } from './crud/connected-accounts';
 import { ContactChannelsCrud } from './crud/contact-channels';
 import { CurrentUserCrud } from './crud/current-user';
@@ -27,7 +28,6 @@ import { TeamInvitationCrud } from './crud/team-invitation';
 import { TeamMemberProfilesCrud } from './crud/team-member-profiles';
 import { TeamPermissionsCrud } from './crud/team-permissions';
 import { TeamsCrud } from './crud/teams';
-import { urlString } from '../utils/urls';
 
 export type ClientInterfaceOptions = {
   clientVersion: string,
@@ -293,6 +293,7 @@ export class StackClientInterface {
         ...(tokenObj?.refreshToken ? {
           "X-Stack-Refresh-Token": tokenObj.refreshToken.token,
         } : {}),
+        "X-Stack-Allow-Anonymous-User": "true",
         ...('publishableClientKey' in this.options ? {
           "X-Stack-Publishable-Client-Key": this.options.publishableClientKey,
         } : {}),
@@ -852,7 +853,7 @@ export class StackClientInterface {
     });
   }
 
-  async signInWithMagicLink(code: string): Promise<Result<{ newUser: boolean, accessToken: string, refreshToken: string }, KnownErrors["VerificationCodeError"]>> {
+  async signInWithMagicLink(code: string, session: InternalSession): Promise<Result<{ newUser: boolean, accessToken: string, refreshToken: string }, KnownErrors["VerificationCodeError"]>> {
     const res = await this.sendClientRequestAndCatchKnownError(
       "/auth/otp/sign-in",
       {
@@ -864,7 +865,7 @@ export class StackClientInterface {
           code,
         }),
       },
-      null,
+      session,
       [KnownErrors.VerificationCodeError]
     );
 
@@ -880,7 +881,7 @@ export class StackClientInterface {
     });
   }
 
-  async signInWithMfa(totp: string, code: string): Promise<Result<{ newUser: boolean, accessToken: string, refreshToken: string }, KnownErrors["VerificationCodeError"]>> {
+  async signInWithMfa(totp: string, code: string, session: InternalSession): Promise<Result<{ newUser: boolean, accessToken: string, refreshToken: string }, KnownErrors["VerificationCodeError"]>> {
     const res = await this.sendClientRequestAndCatchKnownError(
       "/auth/mfa/sign-in",
       {
@@ -894,7 +895,7 @@ export class StackClientInterface {
           code,
         }),
       },
-      null,
+      session,
       [KnownErrors.VerificationCodeError]
     );
 
@@ -910,7 +911,7 @@ export class StackClientInterface {
     });
   }
 
-  async signInWithPasskey(body: { authentication_response: AuthenticationResponseJSON, code: string }): Promise<Result<{accessToken: string, refreshToken: string }, KnownErrors["PasskeyAuthenticationFailed"]>> {
+  async signInWithPasskey(body: { authentication_response: AuthenticationResponseJSON, code: string }, session: InternalSession): Promise<Result<{accessToken: string, refreshToken: string }, KnownErrors["PasskeyAuthenticationFailed"]>> {
     const res = await this.sendClientRequestAndCatchKnownError(
       "/auth/passkey/sign-in",
       {
@@ -920,7 +921,7 @@ export class StackClientInterface {
         },
         body: JSON.stringify(body),
       },
-      null,
+      session,
       [KnownErrors.PasskeyAuthenticationFailed]
     );
 
@@ -945,7 +946,8 @@ export class StackClientInterface {
       state: string,
       type: "authenticate" | "link",
       providerScope?: string,
-    } & ({ type: "authenticate" } | { type: "link", session: InternalSession })
+      session: InternalSession,
+    }
   ): Promise<string> {
     const updatedRedirectUrl = new URL(options.redirectUrl);
     for (const key of ["code", "state"]) {
@@ -972,14 +974,16 @@ export class StackClientInterface {
     url.searchParams.set("type", options.type);
     url.searchParams.set("error_redirect_url", options.errorRedirectUrl);
 
+    const tokens = await options.session.getOrFetchLikelyValidTokens(20_000);
+    if (tokens) {
+      url.searchParams.set("token", tokens.accessToken.token);
+    }
+
     if (options.afterCallbackRedirectUrl) {
       url.searchParams.set("after_callback_redirect_url", options.afterCallbackRedirectUrl);
     }
 
     if (options.type === "link") {
-      const tokens = await options.session.getOrFetchLikelyValidTokens(20_000);
-      url.searchParams.set("token", tokens?.accessToken.token || "");
-
       if (options.providerScope) {
         url.searchParams.set("provider_scope", options.providerScope);
       }
@@ -1750,7 +1754,7 @@ export class StackClientInterface {
     options: (
       { itemId: string, userId: string } |
       { itemId: string, teamId: string } |
-      { itemId: string, customId: string }
+      { itemId: string, customCustomerId: string }
     ),
     session: InternalSession | null,
   ): Promise<ItemCrud['Client']['Read']> {
@@ -1762,11 +1766,11 @@ export class StackClientInterface {
     } else if ("teamId" in options) {
       customerType = "team";
       customerId = options.teamId;
-    } else if ("customId" in options) {
+    } else if ("customCustomerId" in options) {
       customerType = "custom";
-      customerId = options.customId;
+      customerId = options.customCustomerId;
     } else {
-      throw new StackAssertionError("getItem requires one of userId, teamId, or customId");
+      throw new StackAssertionError("getItem requires one of userId, teamId, or customCustomerId");
     }
 
     const response = await this.sendClientRequest(