+++ packages/template/src/lib/stack-app/apps/implementations/client-app-impl.ts
@@ -2,6 +2,7 @@ import { WebAuthnError, startAuthentication, startRegistration } from "@simplewe
 import { KnownErrors, StackClientInterface } from "@stackframe/stack-shared";
 import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
 import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
+import { ItemCrud } from "@stackframe/stack-shared/dist/interface/crud/items";
 import { NotificationPreferenceCrud } from "@stackframe/stack-shared/dist/interface/crud/notification-preferences";
 import { TeamApiKeysCrud, UserApiKeysCrud, teamApiKeysCreateOutputSchema, userApiKeysCreateOutputSchema } from "@stackframe/stack-shared/dist/interface/crud/project-api-keys";
 import { ProjectPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/project-permissions";
@@ -12,7 +13,6 @@ import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/
 import { TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
 import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
 import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
-import { ItemCrud } from "@stackframe/stack-shared/dist/interface/crud/items";
 import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
 import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
 import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
@@ -28,6 +28,8 @@ import { deindent, mergeScopeStrings } from "@stackframe/stack-shared/dist/utils
 import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
 import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
 import * as cookie from "cookie";
+import * as NextNavigationUnscrambled from "next/navigation"; // import the entire module to get around some static compiler warnings emitted by Next.js in some cases | THIS_LINE_PLATFORM next
+import React, { useCallback, useMemo } from "react"; // THIS_LINE_PLATFORM react-like
 import type * as yup from "yup";
 import { constructRedirectUrl } from "../../../../utils/url";
 import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "../../../auth";
@@ -36,18 +38,19 @@ import { ApiKey, ApiKeyCreationOptions, ApiKeyUpdateOptions, apiKeyCreationOptio
 import { GetUserOptions, HandlerUrls, OAuthScopesOnSignIn, RedirectMethod, RedirectToOptions, RequestLike, TokenStoreInit, stackAppInternalsSymbol } from "../../common";
 import { OAuthConnection } from "../../connected-accounts";
 import { ContactChannel, ContactChannelCreateOptions, ContactChannelUpdateOptions, contactChannelCreateOptionsToCrud, contactChannelUpdateOptionsToCrud } from "../../contact-channels";
+import { Customer, InlineOffer, Item } from "../../customers";
 import { NotificationCategory } from "../../notification-categories";
 import { TeamPermission } from "../../permissions";
 import { AdminOwnedProject, AdminProjectUpdateOptions, Project, adminProjectCreateOptionsToCrud } from "../../projects";
 import { EditableTeamMemberProfile, Team, TeamCreateOptions, TeamInvitation, TeamUpdateOptions, TeamUser, teamCreateOptionsToCrud, teamUpdateOptionsToCrud } from "../../teams";
 import { ActiveSession, Auth, BaseUser, CurrentUser, InternalUserExtra, ProjectCurrentUser, UserExtra, UserUpdateOptions, userUpdateOptionsToCrud } from "../../users";
-import { Customer, InlineOffer, Item } from "../../customers";
 import { StackClientApp, StackClientAppConstructorOptions, StackClientAppJson } from "../interfaces/client-app";
 import { _StackAdminAppImplIncomplete } from "./admin-app-impl";
-import { TokenObject, clientVersion, createCache, createCacheBySession, createEmptyTokenStore, getBaseUrl, getDefaultExtraRequestHeaders, getDefaultProjectId, getDefaultPublishableClientKey, getUrls, } from "./common";
-import * as NextNavigationUnscrambled from "next/navigation"; // import the entire module to get around some static compiler warnings emitted by Next.js in some cases | THIS_LINE_PLATFORM next
-import React, { useCallback, useMemo } from "react"; // THIS_LINE_PLATFORM react-like
-import { useAsyncCache } from "./common"; // THIS_LINE_PLATFORM react-like
+import { TokenObject, clientVersion, createCache, createCacheBySession, createEmptyTokenStore, getBaseUrl, getDefaultExtraRequestHeaders, getDefaultProjectId, getDefaultPublishableClientKey, getUrls } from "./common";
+
+// IF_PLATFORM react-like
+import { useAsyncCache } from "./common";
+// END_PLATFORM
 
 let isReactServer = false;
 // IF_PLATFORM next
@@ -214,8 +217,8 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
   );
 
   private readonly _customItemCache = createCacheBySession<[string, string], ItemCrud['Client']['Read']>(
-    async (session, [customId, itemId]) => {
-      return await this._interface.getItem({ customId, itemId }, session);
+    async (session, [customCustomerId, itemId]) => {
+      return await this._interface.getItem({ customCustomerId, itemId }, session);
     }
   );
 
@@ -1204,7 +1207,7 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
     };
   }
 
-  async getItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customId: string }): Promise<Item> {
+  async getItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customCustomerId: string }): Promise<Item> {
     const session = await this._getSession();
     let crud: ItemCrud['Client']['Read'];
     if ("userId" in options) {
@@ -1212,17 +1215,17 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
     } else if ("teamId" in options) {
       crud = Result.orThrow(await this._teamItemCache.getOrWait([session, options.teamId, options.itemId], "write-only"));
     } else {
-      crud = Result.orThrow(await this._customItemCache.getOrWait([session, options.customId, options.itemId], "write-only"));
+      crud = Result.orThrow(await this._customItemCache.getOrWait([session, options.customCustomerId, options.itemId], "write-only"));
     }
     return this._clientItemFromCrud(crud);
   }
 
   // IF_PLATFORM react-like
-  useItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customId: string }): Item {
+  useItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customCustomerId: string }): Item {
     const session = this._useSession();
     const [cache, ownerId] =
       "userId" in options ? [this._userItemCache, options.userId] :
-        "teamId" in options ? [this._teamItemCache, options.teamId] : [this._customItemCache, options.customId];
+        "teamId" in options ? [this._teamItemCache, options.teamId] : [this._customItemCache, options.customCustomerId];
     const crud = useAsyncCache(cache, [session, ownerId, options.itemId] as const, "app.useItem()");
     return this._clientItemFromCrud(crud);
   }
@@ -1444,7 +1447,7 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
     this._ensurePersistentTokenStore(options?.tokenStore);
     const session = await this._getSession(options?.tokenStore);
     let crud = Result.orThrow(await this._currentUserCache.getOrWait([session], "write-only"));
-    if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists") {
+    if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists[deprecated]") {
       crud = null;
     }
 
@@ -1459,10 +1462,10 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
         }
         case 'anonymous': {
           const tokens = await this._signUpAnonymously();
-          return await this.getUser({ tokenStore: tokens, or: "anonymous-if-exists" }) ?? throwErr("Something went wrong while signing up anonymously");
+          return await this.getUser({ tokenStore: tokens, or: "anonymous-if-exists[deprecated]" }) ?? throwErr("Something went wrong while signing up anonymously");
         }
         case undefined:
-        case "anonymous-if-exists":
+        case "anonymous-if-exists[deprecated]":
         case "return-null": {
           return null;
         }
@@ -1482,7 +1485,7 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
 
     const session = this._useSession(options?.tokenStore);
     let crud = useAsyncCache(this._currentUserCache, [session] as const, "useUser()");
-    if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists") {
+    if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists[deprecated]") {
       crud = null;
     }
 
@@ -1509,7 +1512,7 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
           throw new StackAssertionError("suspend should never return");
         }
         case undefined:
-        case "anonymous-if-exists":
+        case "anonymous-if-exists[deprecated]":
         case "return-null": {
           // do nothing
         }
@@ -1534,13 +1537,16 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
     }
 
     this._ensurePersistentTokenStore();
+    const session = await this._getSession();
     await signInWithOAuth(
-      this._interface, {
+      this._interface,
+      {
         provider,
         redirectUrl: this.urls.oauthCallback,
         errorRedirectUrl: this.urls.error,
         providerScope: this._oauthScopesOnSignIn[provider]?.join(" "),
-      }
+      },
+      session,
     );
   }
 
@@ -1653,10 +1659,11 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
 
   async signInWithMagicLink(code: string, options?: { noRedirect?: boolean }): Promise<Result<undefined, KnownErrors["VerificationCodeError"] | KnownErrors["InvalidTotpCode"]>> {
     this._ensurePersistentTokenStore();
+    const session = await this._getSession();
     let result;
     try {
       result = await this._catchMfaRequiredError(async () => {
-        return await this._interface.signInWithMagicLink(code);
+        return await this._interface.signInWithMagicLink(code, session);
       });
     } catch (e) {
       if (KnownErrors.InvalidTotpCode.isInstance(e)) {
@@ -1785,10 +1792,11 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
    */
   async signInWithMfa(totp: string, code: string, options?: { noRedirect?: boolean }): Promise<Result<undefined, KnownErrors["VerificationCodeError"] | KnownErrors["InvalidTotpCode"]>> {
     this._ensurePersistentTokenStore();
+    const session = await this._getSession();
     let result;
     try {
       result = await this._catchMfaRequiredError(async () => {
-        return await this._interface.signInWithMfa(totp, code);
+        return await this._interface.signInWithMfa(totp, code, session);
       });
     } catch (e) {
       if (e instanceof KnownErrors.InvalidTotpCode) {
@@ -1831,7 +1839,7 @@ export class _StackClientAppImplIncomplete<HasTokenStore extends boolean, Projec
         options_json.rpId = window.location.hostname;
 
         const authentication_response = await startAuthentication({ optionsJSON: options_json });
-        return await this._interface.signInWithPasskey({ authentication_response, code });
+        return await this._interface.signInWithPasskey({ authentication_response, code }, session);
       });
     } catch (error) {
       if (error instanceof WebAuthnError) {