+++ packages/template/src/lib/stack-app/apps/implementations/server-app-impl.ts
@@ -33,8 +33,9 @@ import { _StackClientAppImplIncomplete } from "./client-app-impl";
 import { clientVersion, createCache, createCacheBySession, getBaseUrl, getDefaultProjectId, getDefaultPublishableClientKey, getDefaultSecretServerKey } from "./common";
 import { InlineOffer, ServerItem } from "../../customers";
 
-// NEXT_LINE_PLATFORM react-like
+// IF_PLATFORM react-like
 import { useAsyncCache } from "./common";
+// END_PLATFORM
 
 export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, ProjectId extends string> extends _StackClientAppImplIncomplete<HasTokenStore, ProjectId> {
   declare protected _interface: StackServerInterface;
@@ -53,8 +54,9 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
     orderBy?: 'signedUpAt',
     desc?: boolean,
     query?: string,
-  ], UsersCrud['Server']['List']>(async ([cursor, limit, orderBy, desc, query]) => {
-    return await this._interface.listServerUsers({ cursor, limit, orderBy, desc, query });
+    includeAnonymous?: boolean,
+  ], UsersCrud['Server']['List']>(async ([cursor, limit, orderBy, desc, query, includeAnonymous]) => {
+    return await this._interface.listServerUsers({ cursor, limit, orderBy, desc, query, includeAnonymous });
   });
   private readonly _serverUserCache = createCache<string[], UsersCrud['Server']['Read'] | null>(async ([userId]) => {
     const user = await this._interface.getServerUserById(userId);
@@ -170,8 +172,8 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
   );
 
   private readonly _serverCustomItemsCache = createCache<[string, string], ItemCrud['Client']['Read']>(
-    async ([customId, itemId]) => {
-      return await this._interface.getItem({ customId, itemId }, null);
+    async ([customCustomerId, itemId]) => {
+      return await this._interface.getItem({ customCustomerId, itemId }, null);
     }
   );
 
@@ -730,7 +732,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
           ? { itemId: crud.id, userId: customer.id }
           : customer.type === "team"
             ? { itemId: crud.id, teamId: customer.id }
-            : { itemId: crud.id, customId: customer.id };
+            : { itemId: crud.id, customCustomerId: customer.id };
         await app._interface.updateItemQuantity(updateOptions, { delta });
         if (customer.type === "user") await app._serverUserItemsCache.refresh([customer.id, crud.id]);
         else if (customer.type === "team") await app._serverTeamItemsCache.refresh([customer.id, crud.id]);
@@ -741,7 +743,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
           ? { itemId: crud.id, userId: customer.id }
           : customer.type === "team"
             ? { itemId: crud.id, teamId: customer.id }
-            : { itemId: crud.id, customId: customer.id };
+            : { itemId: crud.id, customCustomerId: customer.id };
         await app._interface.updateItemQuantity(updateOptions, { delta: -delta, allow_negative: true });
         if (customer.type === "user") await app._serverUserItemsCache.refresh([customer.id, crud.id]);
         else if (customer.type === "team") await app._serverTeamItemsCache.refresh([customer.id, crud.id]);
@@ -753,7 +755,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
             ? { itemId: crud.id, userId: customer.id }
             : customer.type === "team"
               ? { itemId: crud.id, teamId: customer.id }
-              : { itemId: crud.id, customId: customer.id };
+              : { itemId: crud.id, customCustomerId: customer.id };
           await app._interface.updateItemQuantity(updateOptions, { delta: -delta });
           if (customer.type === "user") await app._serverUserItemsCache.refresh([customer.id, crud.id]);
           else if (customer.type === "team") await app._serverTeamItemsCache.refresh([customer.id, crud.id]);
@@ -846,7 +848,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
       this._ensurePersistentTokenStore(options?.tokenStore);
       const session = await this._getSession(options?.tokenStore);
       let crud = Result.orThrow(await this._currentServerUserCache.getOrWait([session], "write-only"));
-      if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists") {
+      if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists[deprecated]") {
         crud = null;
       }
 
@@ -861,10 +863,10 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
           }
           case 'anonymous': {
             const tokens = await this._signUpAnonymously();
-            return await this.getUser({ tokenStore: tokens, or: "anonymous-if-exists" }) ?? throwErr("Something went wrong while signing up anonymously");
+            return await this.getUser({ tokenStore: tokens, or: "anonymous-if-exists[deprecated]" }) ?? throwErr("Something went wrong while signing up anonymously");
           }
           case undefined:
-          case "anonymous-if-exists":
+          case "anonymous-if-exists[deprecated]":
           case "return-null": {
             return null;
           }
@@ -903,7 +905,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
 
       const session = this._useSession(options?.tokenStore);
       let crud = useAsyncCache(this._currentServerUserCache, [session] as const, "useUser()");
-      if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists") {
+      if (crud?.is_anonymous && options?.or !== "anonymous" && options?.or !== "anonymous-if-exists[deprecated]") {
         crud = null;
       }
 
@@ -930,7 +932,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
             throw new StackAssertionError("suspend should never return");
           }
           case undefined:
-          case "anonymous-if-exists":
+          case "anonymous-if-exists[deprecated]":
           case "return-null": {
             // do nothing
           }
@@ -953,7 +955,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
   // END_PLATFORM
 
   async listUsers(options?: ServerListUsersOptions): Promise<ServerUser[] & { nextCursor: string | null }> {
-    const crud = Result.orThrow(await this._serverUsersCache.getOrWait([options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query], "write-only"));
+    const crud = Result.orThrow(await this._serverUsersCache.getOrWait([options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query, options?.includeAnonymous], "write-only"));
     const result: any = crud.items.map((j) => this._serverUserFromCrud(j));
     result.nextCursor = crud.pagination?.next_cursor ?? null;
     return result as any;
@@ -995,7 +997,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
     return teams.map((t) => this._serverTeamFromCrud(t));
   }
 
-  async getItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customId: string }): Promise<ServerItem> {
+  async getItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customCustomerId: string }): Promise<ServerItem> {
     if ("userId" in options) {
       const result = Result.orThrow(await this._serverUserItemsCache.getOrWait([options.userId, options.itemId], "write-only"));
       return this._serverItemFromCrud({ type: "user", id: options.userId }, result);
@@ -1003,13 +1005,13 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
       const result = Result.orThrow(await this._serverTeamItemsCache.getOrWait([options.teamId, options.itemId], "write-only"));
       return this._serverItemFromCrud({ type: "team", id: options.teamId }, result);
     } else {
-      const result = Result.orThrow(await this._serverCustomItemsCache.getOrWait([options.customId, options.itemId], "write-only"));
-      return this._serverItemFromCrud({ type: "custom", id: options.customId }, result);
+      const result = Result.orThrow(await this._serverCustomItemsCache.getOrWait([options.customCustomerId, options.itemId], "write-only"));
+      return this._serverItemFromCrud({ type: "custom", id: options.customCustomerId }, result);
     }
   }
 
   // IF_PLATFORM react-like
-  useItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customId: string }): ServerItem {
+  useItem(options: { itemId: string, userId: string } | { itemId: string, teamId: string } | { itemId: string, customCustomerId: string }): ServerItem {
     let type: "user" | "team" | "custom";
     let id: string;
     let cache: AsyncCache<[string, string], Result<ItemCrud['Client']['Read']>>;
@@ -1023,7 +1025,7 @@ export class _StackServerAppImplIncomplete<HasTokenStore extends boolean, Projec
       cache = this._serverTeamItemsCache;
     } else {
       type = "custom";
-      id = options.customId;
+      id = options.customCustomerId;
       cache = this._serverCustomItemsCache;
     }