+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/dev-launchpad
 
+## 2.8.34
+
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.32
 
 ## 2.8.31