+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/example-middleware-demo
 
+## 2.8.34
+
+### Patch Changes
+
+- @stackframe/stack@2.8.34
+
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.33
+
 ## 2.8.32
 
 ### Patch Changes