+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # react-example
 
+## 2.8.34
+
+### Patch Changes
+
+- @stackframe/react@2.8.34
+
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/react@2.8.33
+
 ## 2.8.32
 
 ### Patch Changes